# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "description": "Documentation officielle de l'API Shop",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Lister toutes les catégories",
                "description": "Retourne toutes les catégories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requête traitée avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Informatique"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une catégorie",
                "description": "Création d'une nouvelle catégorie",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom unique de la catégorie",
                                        "type": "string",
                                        "example": "Robotique"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "La catégorie a bien été créée"
                    },
                    "409": {
                        "description": "Cette catégorie existe déjà"
                    },
                    "500": {
                        "description": "Une erreur est survenue ."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/categories/{category}": {
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Modifier une catégorie",
                "description": "Modifier le nom d'une catégorie",
                "operationId": "02aa39951e67916277e28086f85909c8",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Robotique Industrielle"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie mise à jour avec succès"
                    },
                    "409": {
                        "description": "Cette catégorie existe déjà"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie",
                "description": "Supprime une catégorie si elle existe",
                "operationId": "d9ecd5c4a74c4d75339251863eedc08c",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée avec succès"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Catégories",
            "description": "Gestion des catégories"
        }
    ]
}